first part of correct bugs
