+++ database/factories/TopicFactory.php
@@ -17,7 +17,7 @@ class TopicFactory extends Factory
     public function definition(): array
     {
         return [
-            'name' => $this->faker->unique()->word,
+            'name' => $this->faker->unique()->word(),
             'created_at' => now(),
             'updated_at' => now(),
         ];
