+++ tests/Feature/Auth/PasswordUpdateTest.php
@@ -13,7 +13,7 @@ class PasswordUpdateTest extends TestCase
 
     public function test_password_can_be_updated(): void
     {
-        $user = User::factory()->create();
+        $user = User::factory()->create($this->dataLogin());
 
         $response = $this
             ->actingAs($user)
@@ -33,7 +33,7 @@ class PasswordUpdateTest extends TestCase
 
     public function test_correct_password_must_be_provided_to_update_password(): void
     {
-        $user = User::factory()->create();
+        $user = User::factory()->create($this->dataLogin());
 
         $response = $this
             ->actingAs($user)
@@ -48,4 +48,13 @@ class PasswordUpdateTest extends TestCase
             ->assertSessionHasErrorsIn('updatePassword', 'current_password')
             ->assertRedirect('/profile');
     }
+
+    private function dataLogin()
+        {
+            return [
+                'email' => 'users@user.com',
+                'password' => 'password',
+                'email_verified_at' => now(),
+            ];
+        }
 }
