--- tests/Feature/EmailVerificationTest.php
@@ -1,72 +1,0 @@
-<?php
-
-namespace Tests\Feature\Auth;
-
-use App\Models\User;
-use App\Providers\RouteServiceProvider;
-use Illuminate\Auth\Events\Verified;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Facades\Event;
-use Illuminate\Support\Facades\URL;
-use Tests\TestCase;
-
-class EmailVerificationTest extends TestCase
-{
-    use RefreshDatabase;
-
-    public function test_email_verification(): void
-    {
-        $user = User::factory()->create([
-            'email_verified_at' => null,
-        ]);
-
-        $response = $this->actingAs($user)->get(route('verification.verify', ['id' => $user->id, 'hash' => sha1($user->email)]));
-
-        $this->assertTrue($user->fresh()->hasVerifiedEmail());
-        $response->assertJson([
-            'message' => __('Email has been verified'),
-            'data' => [
-                'verified' => true,
-                'link_verification' => false,
-                'already_verified' => false,
-            ],
-        ]);
-    }
-
-    public function test_resend_verification_email(): void
-    {
-        $user = User::factory()->create([
-            'email_verified_at' => null,
-        ]);
-
-        $response = $this->actingAs($user)->post(route('verification.resend'));
-
-        $this->assertNull($user->fresh()->email_verified_at);
-
-        $response->assertJson([
-            'message' => __('New email verification link sent successfully.'),
-            'data' => [
-                'verified' => false,
-                'link_verification' => true,
-                'already_verified' => false,
-            ],
-        ]);
-    }
-
-    public function test_email_is_not_verified_with_invalid_hash(): void
-    {
-        $user = User::factory()->create([
-            'email_verified_at' => null,
-        ]);
-
-        $verificationUrl = URL::temporarySignedRoute(
-            'verification.verify',
-            now()->addMinutes(60),
-            ['id' => $user->id, 'hash' => sha1('wrong-email')]
-        );
-
-        $this->actingAs($user)->get($verificationUrl);
-
-        $this->assertFalse($user->fresh()->hasVerifiedEmail());
-    }
-}
